+++ tsrc/modules/firebase.ts
@@ -1,0 +1,2 @@
+import { firebaseConfig } from "./../firebase";
+import { initializeApp } from "firebase/app";
