+++ tsrc/modules/hub.ts
@@ -1,14 +1,21 @@
 import { Module } from "../moduleBase";
-import type { GameState, AppSettings, HubReceive } from "../utility";
-import { MicroLobbyData } from "wc3mt-lobby-container";
+import type { GameState } from "../utility";
+import type { MicroLobbyData } from "wc3mt-lobby-container";
+
+import { WebSocket } from "ws";
+import { HubReceive, GameStateUpdate, SettingsUpdate } from "../utility";
+import type { LobbyUpdates } from "wc3mt-lobby-container";
+
+import { settings } from "./../globals/settings";
 
 export class HubControl extends Module {
   hubWebSocket: WebSocket | null = null;
   isPackaged: boolean;
   appVersion: string;
+  #heartBeatTimer: NodeJS.Timeout | null = null;
+
   constructor(
     baseModule: {
-      settings: AppSettings;
       gameState: GameState;
       identifier: string;
       lobby?: MicroLobbyData;
@@ -22,8 +29,40 @@ export class HubControl extends Module {
     this.socketSetup();
   }
 
-  socketSetup() {
-    if (!this.hubWebSocket) {
+  updateGameState(updates: GameStateUpdate): boolean {
+    let updated = super.updateGameState(updates);
+    if (updated) {
+      this.sendToHub({
+        gameStateUpdate: updates,
+      });
+    }
+    return updated;
+  }
+
+  updateSettings(updates: SettingsUpdate): boolean {
+    let updated = super.updateSettings(updates);
+    if (updated) {
+      this.sendToHub({
+        settingsUpdate: updates,
+      });
+    }
+    return updated;
+  }
+
+  updateLobby(update: LobbyUpdates): { isUpdated: boolean; events: LobbyUpdates[] } {
+    let updates = super.updateLobby(update);
+    if (updates.isUpdated) {
+      this.sendToHub({
+        lobbyUpdates: update,
+      });
+    }
+    return updates;
+  }
+
+  private socketSetup() {
+    if (this.#heartBeatTimer) {
+      clearTimeout(this.#heartBeatTimer);
+      this.#heartBeatTimer = null;
     }
     if (this.isPackaged) {
       this.hubWebSocket = new WebSocket("wss://ws.trenchguns.com/" + this.identifier);
@@ -31,40 +70,40 @@ export class HubControl extends Module {
       this.hubWebSocket = new WebSocket("wss://wsdev.trenchguns.com/" + this.identifier);
     }
     this.hubWebSocket.onerror = (error) => {
-      if (app.isPackaged) this.emitError("Failed hub connection: " + error);
+      if (this.isPackaged) this.emitError("Failed hub connection: " + error);
     };
     this.hubWebSocket.onopen = (ev) => {
       if (this.hubWebSocket?.readyState !== WebSocket.OPEN) return;
       this.emitInfo("Connected to hub");
-      if (
-        this.lobby?.lobbyStatic &&
-        (!this.settings.autoHost.private || !app.isPackaged)
-      ) {
-        this.sendToHub("lobbyUpdate", {
+      if (this.lobby?.lobbyStatic && (!settings.autoHost.private || !this.isPackaged)) {
+        this.sendToHub({
           lobbyUpdates: { newLobby: this.lobby.exportMin() },
         });
       }
-      setTimeout(this.hubHeartbeat, 30000);
+      this.#heartBeatTimer = setInterval(this.hubHeartbeat, 30000);
     };
     this.hubWebSocket.onmessage = (data) => {
       this.emitInfo("Received message from hub: " + data);
     };
     this.hubWebSocket.onclose = (ev) => {
-      if (app.isPackaged) this.emitError("Disconnected from hub");
+      if (this.isPackaged) this.emitError("Disconnected from hub");
       setTimeout(this.socketSetup, Math.random() * 5000 + 3000);
+      if (this.#heartBeatTimer) {
+        clearTimeout(this.#heartBeatTimer);
+        this.#heartBeatTimer = null;
+      }
       this.hubWebSocket = null;
     };
   }
 
-  hubHeartbeat() {
+  private hubHeartbeat() {
     if (this.hubWebSocket?.OPEN) {
-      this.sendToHub("heartbeat");
-      setTimeout(this.hubHeartbeat, 30000);
+      this.sendToHub({ heartbeat: true });
     }
   }
 
-  sendToHub(messageType: HubReceive["messageType"], data?: HubReceive["data"]) {
-    let buildMessage: HubReceive = { messageType, data, appVersion: this.appVersion };
+  sendToHub(data: HubReceive["data"]) {
+    let buildMessage: HubReceive = { data, appVersion: this.appVersion };
     if (this.hubWebSocket && this.hubWebSocket.readyState === WebSocket.OPEN) {
       this.hubWebSocket.send(JSON.stringify(buildMessage));
     }
